Fixed warnings; refs #3134
git-svn-id: http://svn.boost.org/svn/boost/trunk@53932 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ libs/graph_parallel/example/dijkstra_shortest_paths.cpp
@@ -78,7 +78,7 @@ int main(int argc, char* argv[])
   if (argc > 2)
     outfile = argv[2];
   else {
-    int i = outfile.rfind('.');
+    size_t i = outfile.rfind('.');
     if (i != std::string::npos)
       outfile.erase(outfile.begin() + i, outfile.end());
     outfile += "-dijkstra.dot";
